improved build script

+++ build.sh
@@ -13,11 +13,11 @@ export UNIONFS_SIZE="2G"
 export APKLIST="$(sh $PROFILEDIR/apklist.sh)"
 export REPODIR="$PROFILEDIR/repo"
 export BUILD_CMD="sudo -E sh build-minirootfs.sh"
-export ROOTFS_URL="https://dl-cdn.alpinelinux.org/alpine/v3.16/releases/$PROFILEARCH/alpine-minirootfs-3.16.2-$PROFILEARCH.tar.gz"
-export ROOTFS_APKS="alpine-base openrc busybox-initscripts busybox kbd-bkeymaps chrony dhcpcd e2fsprogs haveged network-extras openntpd openssl openssh tzdata wget sigma-rootfs"
+export ROOTFS_APKS="alpine-base openrc busybox-initscripts busybox kbd-bkeymaps chrony dhcpcd e2fsprogs haveged network-extras openntpd openssl openssh tzdata wget"
 export INITFS_FEATURES="ata base bootchart cdrom ext4 mmc nvme raid scsi squashfs usb virtio"
 export KERNEL_FLAVOR="lts"
 export ZSTD_LEVEL="22"
+export IGNORE_SIGMA_ROOTFS="no"
 
 # parse args
 getval() {
@@ -30,15 +30,20 @@ for arg in $@; do
         "-q") export ZSTD_LEVEL="9";;
 
         # custom arch
-        "-a="*) export PROFILEARCH=$(getval $arg);;
+        "-a="*) export PROFILEARCH="$(getval $arg)";;
 
         # custom kernel flavor
         "-k="*) export KERNEL_FLAVOR="$(getval $arg)";;
 
         # legacy build
         "-l") export BUILD_CMD="sh build-mkimg.sh";;
+
+        # ignore sigma-rootfs (faster build for testing)
+        "-i") export IGNORE_SIGMA_ROOTFS="yes";;
     esac
 done
+
+export ROOTFS_URL="https://dl-cdn.alpinelinux.org/alpine/v3.16/releases/$PROFILEARCH/alpine-minirootfs-3.16.2-$PROFILEARCH.tar.gz"
 
 # show build info
 echo "---------------------"
@@ -48,33 +53,41 @@ echo "Profile: $PROFILENAME"
 echo "Arch: $PROFILEARCH"
 echo "Kernel: linux-$KERNEL_FLAVOR"
 echo "Zstd level: $ZSTD_LEVEL"
+echo "Ignore Sigma RootFS? $IGNORE_SIGMA_ROOTFS"
 echo "Build command: $BUILD_CMD"
-
-exit 0
+echo "---------------------"
 
 # create required directories
+echo "[*] Setting up directories..."
 mkdir -p "$CACHEDIR" "$OUTDIR"
 
-# create user directories
-for dir in Downloads Documents Pictures Videos Music; do
-	mkdir -p "$PROFILEDIR/apk/sigma-rootfs/rootfs/etc/skel/$dir"
-done
+if [ "$IGNORE_SIGMA_ROOTFS" != "yes" ]; then
+    export ROOTFS_APKS="$ROOTFS_APKS sigma-rootfs"
+    # create user directories
+    for dir in Downloads Documents Pictures Videos Music; do
+    	mkdir -p "$PROFILEDIR/apk/sigma-rootfs/rootfs/etc/skel/$dir"
+    done
 
-# build sigma rootfs apk
-mkdir -p "$REPODIR"
-cd "$APKDIR"/sigma-rootfs
-tar -zcf rootfs.tar.gz rootfs
-export ROOTFS_SHA512="$(sha512sum rootfs.tar.gz)"
-abuild -rf -P "$REPODIR"
-rm rootfs.tar.gz
-cd "$BASEDIR"
+    # build sigma rootfs apk
+    echo "[*] Creating 'sigma-rootfs'..."
+    mkdir -p "$REPODIR"
+    cd "$APKDIR"/sigma-rootfs
+    tar -zcf rootfs.tar.gz rootfs
+    export ROOTFS_SHA512="$(sha512sum rootfs.tar.gz)"
+    abuild -rf -P "$REPODIR"
+    rm rootfs.tar.gz
+    cd "$BASEDIR"
+fi
 
 # build iso
+echo "[*] Starting build command: $BUILD_CMD"
 sudo -E sh build-minirootfs.sh
-# sh build-mkimg.sh
+echo "[*] Build complete"
 
 # create checksum
 cd "$OUTDIR"
 sha256sum *.iso > sha256sums
 cd "$BASEDIR"
+echo "[*] Output files at: $OUTDIR"
+
